sweetalert use in add,edit,remove section

+++ src/redux/thunk/content/removeContent.js
@@ -1,3 +1,4 @@
+import Swal from "sweetalert2"
 import { remove_content } from "../../action/postAction"
 
 const removeContent = (id)=>{
@@ -6,7 +7,20 @@ const removeContent = (id)=>{
             method: 'DELETE',
         })
         const data = await res.json()
-        dispatch(remove_content(id))
+        if(data.acknowledged){
+            Swal.fire({
+                title:"Success",
+                text:"Post Delete Successfully!",
+                icon:"success"
+            })
+            dispatch(remove_content(id))
+        }else{
+            Swal.fire({
+                title:"Oops",
+                text:"Post Delete Failed!",
+                icon:"error"
+            })
+        }
     }
 }
 export default removeContent
